requirements.txt
python-telegram-bot[job-queue]==21.6
asyncpg
aiohttp==3.10.5
APScheduler==3.10.4
pytz==2024.2
python-dotenv==1.0.1
Pillow>=10.0.0
